socket enabled and api calls
