+++ package.json
@@ -32,6 +32,7 @@
     "bluebird": "^3.4.6",
     "body-parser": "^1.15.2",
     "browser-sync": "^2.17.3",
+    "classnames": "^2.2.5",
     "deepcopy": "^0.6.1",
     "express": "^4.14.0",
     "express-promise": "^0.4.0",
