+++ src/public/scss/calendar.scss
@@ -132,6 +132,10 @@
       &.full {
         opacity: 0.5;
       }
+
+      &.anchored {
+        border: 2px solid $red1;
+      }
     }
 
     &#A > li.event {
