+++ src/public/views/Scheduler.js
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import _ from 'lodash';
 
-import CourseList from '../modules/CourseList';
-import Calendar from '../modules/Calendar';
-import SelectorFilter from '../modules/SelectorFilter';
+import CourseList from '../containers/CourseList';
+import Calendar from '../containers/Calendar';
+import SelectorFilter from '../containers/SelectorFilter';
 
 import ApiInterface from '../api-interface';
 let api = new ApiInterface();
@@ -12,14 +12,21 @@ class Scheduler extends Component {
 
   constructor(props) {
     super(props);
+    let uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
+      var r = Math.random()*16|0, v = c == 'x' ? r : (r&0x3|0x8);
+      return v.toString(16);
+    });
     this.state = {
+      uuid: uuid,
       courses: [],
       courseData: {},
       combinations: [],
+      anchors: {},
       colors: [],
-      index: null,
+      index: 0,
       hover: null,
     };
+    this.socket = io();
   }
 
   render() {
@@ -41,6 +48,8 @@ class Scheduler extends Component {
           index={this.state.index}
           hoverIndex={this.state.hover}
           setHover={this.setHover.bind(this)}
+          anchors={this.state.anchors}
+          toggleAnchor={this.toggleAnchor.bind(this)}
           colors={this.state.colors} />
         <SelectorFilter
           courses={this.state.courses}
@@ -55,14 +64,20 @@ class Scheduler extends Component {
   componentDidMount() {
     document.addEventListener('keydown', this.keyboardCommands.bind(this), false);
     document.getElementById('search').addEventListener('keydown', this.keyboardCommandsInInput.bind(this), false);
+
+    // sockets
+    this.socket.on('receive:courseData', function (courseId) {
+      if (this.state.courses.indexOf(courseId) > -1) {
+        this.generateSchedules();
+      }
+    });
   }
 
   addClass(courseId) {
     let _this = this;
 
-    if (this.state.courses.indexOf(courseId) > -1) {
-      return;
-    }
+    // don't re-add class
+    if (this.state.courses.indexOf(courseId) > -1) return;
 
     // 1. verify that the course exists
     api.verify(courseId, 20163).then(courseExists => {
@@ -77,9 +92,47 @@ class Scheduler extends Component {
   }
 
   removeClass(courseId) {
-    this.setState({ courses: _.pull(this.state.courses, courseId) }, () => {
-      this.generateSchedules();
-    });
+    let anchors = this.state.anchors;
+    if (anchors[courseId]) {
+      delete anchors[courseId];
+      this.setState({ anchors });
+    }
+
+    this.setState({
+      courses: _.pull(this.state.courses, courseId)
+    }, this.generateSchedules);
+  }
+
+  toggleAnchor(courseId, sectionId) {
+    let anchors = this.state.anchors;
+    if (anchors[courseId] && anchors[courseId].indexOf(sectionId) >= 0) {
+      this.removeAnchor(courseId, sectionId);
+    } else {
+      this.addAnchor(courseId, sectionId);
+    }
+  }
+
+  addAnchor(courseId, sectionId) {
+    let anchors = this.state.anchors;
+
+    if (!_.isArray(anchors[courseId]))
+      anchors[courseId] = [];
+
+    if (anchors[courseId].indexOf(sectionId) >= 0) return;
+
+    anchors[courseId].push(sectionId);
+    this.setState({ anchors }, this.generateSchedules);
+  }
+
+  removeAnchor(courseId, sectionId) {
+    let anchors = this.state.anchors;
+    if (!anchors[courseId]) return;
+    if (anchors[courseId].indexOf(sectionId) < 0) return;
+    anchors[courseId] = _.pull(anchors[courseId], sectionId)
+    if (anchors[courseId].length <= 0) {
+      delete anchors[courseId];
+    }
+    this.setState({ anchors }, this.generateSchedules);
   }
 
   generateSchedules() {
@@ -91,12 +144,17 @@ class Scheduler extends Component {
         index: 0,
       });
     } else {
-      api.generateSchedules(this.state.courses).then(({ courseData, results }) => {
-        this.setState({
-          courseData,
-          combinations: results,
-          index: 0
-        });
+      api.generateCourseDataAndSchedules(this.state.courses, this.state.anchors)
+        .then(({ courseData, results }) => {
+          console.log(courseData, results);
+          let index = this.state.index;
+          if (index >= results.length) index = results.length - 1;
+          if (index <= 0) index = 0;
+          this.setState({
+            courseData,
+            combinations: results,
+            index
+          });
       });
     }
   }
@@ -106,24 +164,6 @@ class Scheduler extends Component {
   }
 
   generateColors() {
-    // var colors = [];
-    // let count = this.state.courses.length;
-    // for (var i = 0; i < count; i++) {
-    //   colors.push(colorFade([233,52,50],[233,167,30], i, count));
-    // }
-    // this.setState({ colors });
-    //
-    // function colorFade(startColor, endColor, index, count){
-    //   var diffR = endColor[0] - startColor[0],
-    //       diffG = endColor[1] - startColor[1],
-    //       diffB = endColor[2] - startColor[2],
-    //       percentFade = (index + 1) / count;
-    //   diffR = Math.round((diffR * percentFade) + startColor[0]);
-    //   diffG = Math.round((diffG * percentFade) + startColor[1]);
-    //   diffB = Math.round((diffB * percentFade) + startColor[2]);
-    //   return [diffR, diffG, diffB];
-    // }
-
     function rgb(r,g,b) {
       return [r,g,b];
     }
