+++ www/index.html
@@ -6,11 +6,10 @@
   <meta http-equiv='x-ua-compatible' content='ie=edge' />
   <link rel='stylesheet' type='text/css' href='res/main.css'>
   <title>HackSChedule</title>
-  <script src="https://use.typekit.net/eex3uxh.js"></script>
-  <script>try{Typekit.load({ async: true });}catch(e){}</script>
 </head>
 <body>
   <div id="app"></div>
+  <script src="/socket.io/socket.io.js"></script>
   <script src="/res/app.js"></script>
 </body>
 </html>
